Convert interceptor calling to template pattern.

+++ resource-terminator/src/main/java/io/github/odalabasmaz/awsgenie/terminator/IAMPolicyResourceTerminator.java
@@ -10,23 +10,20 @@ import io.github.odalabasmaz.awsgenie.fetcher.credentials.AWSClientConfiguration
 import io.github.odalabasmaz.awsgenie.fetcher.credentials.AWSClientProvider;
 import io.github.odalabasmaz.awsgenie.fetcher.iam.IAMPolicyResource;
 import io.github.odalabasmaz.awsgenie.terminator.configuration.Configuration;
-import io.github.odalabasmaz.awsgenie.terminator.interceptor.InterceptorRegistry;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import java.text.SimpleDateFormat;
-import java.util.Date;
-import java.util.LinkedHashSet;
-import java.util.LinkedList;
-import java.util.List;
+import java.util.*;
 import java.util.concurrent.TimeUnit;
+import java.util.stream.Collectors;
 
 /**
  * @author Orhun Dalabasmaz
  * @version 06.04.2021
  */
 
-public class IAMPolicyResourceTerminator extends ResourceTerminatorWithProvider implements ResourceTerminator<IAMPolicyResource> {
+public class IAMPolicyResourceTerminator extends ResourceTerminator<IAMPolicyResource> {
     private static final Logger LOGGER = LogManager.getLogger(IAMPolicyResourceTerminator.class);
 
     private ResourceFetcherFactory<IAMPolicyResource> resourceFetcherFactory;
@@ -36,24 +33,20 @@ public class IAMPolicyResourceTerminator extends ResourceTerminatorWithProvider 
     }
 
     @Override
-    public void terminateResource(Configuration conf, boolean apply) throws Exception {
-        terminateResource(conf.getRegion(), Service.fromValue(conf.getService()), conf.getResourcesAsList(), conf.getDescription(),
+    public Set<IAMPolicyResource> beforeApply(Configuration conf, boolean apply) throws Exception {
+        return beforeApply(conf.getRegion(), Service.fromValue(conf.getService()), conf.getResourcesAsList(), conf.getDescription(),
                 conf.getLastUsage(), conf.isForce(), apply);
     }
 
     @Override
-    public void terminateResource(String region, Service service, List<String> resources, String ticket, boolean apply) throws Exception {
-        terminateResource(region, service, resources, ticket, 7, false, apply);
+    protected Set<IAMPolicyResource> beforeApply(String region, Service service, List<String> resources, String ticket, boolean apply) throws Exception {
+        return beforeApply(region, service, resources, ticket, 7, false, apply);
     }
 
     @Override
-    public void terminateResource(String region, Service service, List<String> resources, String ticket, int lastUsage, boolean force, boolean apply) throws Exception {
-
-        AmazonIdentityManagement iamClient = AWSClientProvider.getInstance(getConfiguration()).getAmazonIAM();
-
+    protected Set<IAMPolicyResource> beforeApply(String region, Service service, List<String> resources, String ticket, int lastUsage, boolean force, boolean apply) throws Exception {
         // Resources to be removed
-        LinkedHashSet<String> policiesToDelete = new LinkedHashSet<>();
-
+        Set<IAMPolicyResource> policiesToDelete = new LinkedHashSet<>();
         List<String> details = new LinkedList<>();
 
         SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
@@ -61,12 +54,12 @@ public class IAMPolicyResourceTerminator extends ResourceTerminatorWithProvider 
         Date referenceDate = new Date(endDate.getTime() - TimeUnit.DAYS.toMillis(lastUsage));
 
         ResourceFetcher<IAMPolicyResource> fetcher = getFetchResourceFactory().getFetcher(service, new ResourceFetcherConfiguration(getConfiguration()));
-        List<IAMPolicyResource> iamPolicyResourceList = fetcher.fetchResources(region, resources, details);
+        Set<IAMPolicyResource> iamPolicyResourceList = fetcher.fetchResources(region, resources, details);
 
         for (IAMPolicyResource iamPolicyResource : iamPolicyResourceList) {
             String policyName = iamPolicyResource.getResourceName();
-            //TODO: fetching later, why not fetched at first...
             Date lastUsedDate = (Date) fetcher.getUsage(region, policyName, lastUsage);
+
             if (lastUsedDate != null && lastUsedDate.after(referenceDate)) {
                 if (force) {
                     details.add("IAM policy seems in use, but still deleting with force: [" + policyName + "], lastUsageDate: [" + sdf.format(lastUsedDate) + "]");
@@ -77,7 +70,8 @@ public class IAMPolicyResourceTerminator extends ResourceTerminatorWithProvider 
                     continue;
                 }
             }
-            policiesToDelete.add(policyName);
+
+            policiesToDelete.add(iamPolicyResource);
         }
 
         StringBuilder info = new StringBuilder()
@@ -89,17 +83,26 @@ public class IAMPolicyResourceTerminator extends ResourceTerminatorWithProvider 
         details.forEach(d -> info.append("-- ").append(d).append("\n"));
         LOGGER.info(info);
 
-        InterceptorRegistry.getBeforeTerminateInterceptors()
-                .forEach(interceptor -> interceptor.intercept(service, iamPolicyResourceList, info.toString(), apply));
+        return policiesToDelete;
+    }
 
-        if (apply) {
+    @Override
+    protected void apply(Set<IAMPolicyResource> resources, boolean apply) {
+        if (!resources.isEmpty() && apply) {
+            AmazonIdentityManagement iamClient = AWSClientProvider.getInstance(getConfiguration()).getAmazonIAM();
+
+            List<String> policiesToDelete = resources
+                    .stream()
+                    .map(IAMPolicyResource::getResourceName)
+                    .collect(Collectors.toList());
+
             LOGGER.info("Terminating the resources...");
             policiesToDelete.forEach(r -> iamClient.deletePolicy(new DeletePolicyRequest().withPolicyArn(r)));
         }
+    }
 
-        InterceptorRegistry.getAfterTerminateInterceptors()
-                .forEach(interceptor -> interceptor.intercept(service, iamPolicyResourceList, info.toString(), apply));
-
+    @Override
+    protected void afterApply(Set<IAMPolicyResource> resources) {
         LOGGER.info("Succeed.");
     }
